Moved App.js file out of components folder and into src folder
