+++ smurfs/src/App.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import { getSmurfs } from '../actions';
+import { getSmurfs } from './actions';
 import './App.css';
 /*
  to wire this component up you're going to need a few things.
