+++ smurfs/src/components/SmurfsList.js
@@ -1,0 +1,2 @@
+import React from 'react';
+
